refactor: 경매 환불을 Auctioneer에서 담당하도록 수정
- 기존 PaymentService에 존재하던 경매 환불 기능을 제거한다.
